water_resource_system/data_acquisition/scheduler.py
"""
数据采集调度系统
基于Apache Airflow的自动化调度
"""
import logging
from datetime import datetime, timedelta
from pathlib import Path
from typing import Dict, List, Any
from ..config import API_CONFIGS, CACHE_DIR

logger = logging.getLogger(__name__)


class DataScheduler:
    """数据采集调度器"""
    
    def __init__(self):
        self.collectors = {}
        self.schedule_config = {
            "remote_sensing": {
                "frequency": "daily",
                "time": "02:00",
                "lookback_days": 3,
            },
            "meteorology": {
                "frequency": "every_6_hours",
                "time": ["00:00", "06:00", "12:00", "18:00"],
                "lookback_days": 1,
            },
            "hydrology": {
                "frequency": "hourly",
                "lookback_days": 1,
            },
        }
    
    def register_collector(self, name: str, collector):
        """
        注册数据采集器
        
        Args:
            name: 采集器名称
            collector: 采集器实例
        """
        self.collectors[name] = collector
        logger.info(f"注册采集器: {name}")
    
    def create_airflow_dag(self) -> str:
        """
        创建Airflow DAG配置
        
        Returns:
            DAG Python代码
        """
        dag_code = '''
"""
水资源监测数据采集DAG
自动化数据更新工作流
"""
from datetime import datetime, timedelta
from airflow import DAG
from airflow.operators.python import PythonOperator
from airflow.operators.bash import BashOperator

# 默认参数
default_args = {
    'owner': 'water_resource_team',
    'depends_on_past': False,
    'start_date': datetime(2025, 1, 1),
    'email': ['alert@example.com'],
    'email_on_failure': True,
    'email_on_retry': False,
    'retries': 2,
    'retry_delay': timedelta(minutes=5),
}

# 创建DAG
dag = DAG(
    'water_resource_data_collection',
    default_args=default_args,
    description='自动化水资源数据采集',
    schedule_interval='0 2 * * *',  # 每天凌晨2点运行
    catchup=False,
    tags=['water_resource', 'data_collection'],
)


def collect_remote_sensing_data(**context):
    """采集遥感数据"""
    from water_resource_system.data_acquisition.remote_sensing.sentinel import SentinelCollector
    from water_resource_system.config import API_CONFIGS, CACHE_DIR
    
    # 创建采集器
    collector = SentinelCollector(
        API_CONFIGS['sentinel'],
        CACHE_DIR / 'sentinel'
    )
    
    # 采集最近3天的数据
    end_date = datetime.now()
    start_date = end_date - timedelta(days=3)
    
    files = collector.collect(start_date, end_date)
    print(f"采集到 {len(files)} 个遥感数据文件")
    
    return len(files)


def collect_meteorology_data(**context):
    """采集气象数据"""
    from water_resource_system.data_acquisition.meteorology.ecmwf import ECMWFCollector
    from water_resource_system.config import API_CONFIGS, CACHE_DIR
    
    collector = ECMWFCollector(
        API_CONFIGS['ecmwf'],
        CACHE_DIR / 'ecmwf'
    )
    
    # 采集昨天的数据
    end_date = datetime.now() - timedelta(days=1)
    start_date = end_date
    
    files = collector.collect(start_date, end_date)
    print(f"采集到 {len(files)} 个气象数据文件")
    
    return len(files)


def collect_forecast_data(**context):
    """采集预报数据"""
    from water_resource_system.data_acquisition.meteorology.ecmwf import ECMWFCollector
    from water_resource_system.config import API_CONFIGS, CACHE_DIR
    
    collector = ECMWFCollector(
        API_CONFIGS['ecmwf'],
        CACHE_DIR / 'ecmwf'
    )
    
    # 采集最新预报
    forecast_date = datetime.now()
    file_path = collector.collect_forecast(forecast_date)
    
    if file_path:
        print(f"预报数据采集完成: {file_path}")
        return True
    return False


def process_water_extraction(**context):
    """处理水体提取"""
    print("开始水体提取处理...")
    # 这里调用水体提取算法
    return True


def run_hydrological_model(**context):
    """运行水文模型"""
    print("开始运行水文模型...")
    # 这里调用水文模型计算
    return True


def generate_forecast(**context):
    """生成预报结果"""
    print("生成水资源预报...")
    # 这里生成预报产品
    return True


def send_alert(**context):
    """发送预警信息"""
    print("检查预警条件并发送通知...")
    # 这里检查是否需要预警
    return True


# 定义任务
task_remote_sensing = PythonOperator(
    task_id='collect_remote_sensing',
    python_callable=collect_remote_sensing_data,
    dag=dag,
)

task_meteorology = PythonOperator(
    task_id='collect_meteorology',
    python_callable=collect_meteorology_data,
    dag=dag,
)

task_forecast = PythonOperator(
    task_id='collect_forecast',
    python_callable=collect_forecast_data,
    dag=dag,
)

task_water_extraction = PythonOperator(
    task_id='process_water_extraction',
    python_callable=process_water_extraction,
    dag=dag,
)

task_model_run = PythonOperator(
    task_id='run_hydrological_model',
    python_callable=run_hydrological_model,
    dag=dag,
)

task_generate_forecast = PythonOperator(
    task_id='generate_forecast',
    python_callable=generate_forecast,
    dag=dag,
)

task_alert = PythonOperator(
    task_id='send_alert',
    python_callable=send_alert,
    dag=dag,
)

# 定义任务依赖关系
task_remote_sensing >> task_water_extraction
task_meteorology >> task_model_run
task_forecast >> task_model_run
task_water_extraction >> task_model_run
task_model_run >> task_generate_forecast
task_generate_forecast >> task_alert
'''
        return dag_code
    
    def save_airflow_dag(self, output_path: Path):
        """
        保存Airflow DAG到文件
        
        Args:
            output_path: 输出路径
        """
        dag_code = self.create_airflow_dag()
        with open(output_path, 'w', encoding='utf-8') as f:
            f.write(dag_code)
        
        logger.info(f"Airflow DAG已保存到: {output_path}")
    
    def run_manual_collection(self, collector_name: str,
                             start_date: datetime,
                             end_date: datetime,
                             region: Dict = None) -> List:
        """
        手动运行数据采集
        
        Args:
            collector_name: 采集器名称
            start_date: 开始日期
            end_date: 结束日期
            region: 区域范围
            
        Returns:
            采集结果
        """
        if collector_name not in self.collectors:
            logger.error(f"采集器不存在: {collector_name}")
            return []
        
        collector = self.collectors[collector_name]
        
        logger.info(f"开始手动采集: {collector_name}")
        logger.info(f"时间范围: {start_date} 到 {end_date}")
        
        try:
            results = collector.collect(start_date, end_date, region)
            logger.info(f"采集完成，共 {len(results)} 个文件")
            return results
        except Exception as e:
            logger.error(f"采集失败: {e}")
            return []
