requirements.txt
# 中国水库自动化水资源监测与预测系统 - 依赖包

# 核心数据处理
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0
xarray>=0.19.0
netCDF4>=1.5.7

# 遥感数据处理
GDAL>=3.3.0
rasterio>=1.2.0
sentinelsat>=1.1.0
landsatxplore>=0.13.0
modis-tools>=1.1.0

# 气象数据
cdsapi>=0.5.1
cfgrib>=0.9.9
herbie-data>=0.0.9

# 地理空间处理
geopandas>=0.10.0
shapely>=1.8.0
fiona>=1.8.0
pyproj>=3.2.0

# 数据库
psycopg2-binary>=2.9.0
influxdb>=5.3.1
redis>=4.0.0

# Web框架
fastapi>=0.85.0
uvicorn>=0.18.0
pydantic>=1.9.0

# 任务调度
apache-airflow>=2.4.0
celery>=5.2.0

# 机器学习
scikit-learn>=1.0.0
tensorflow>=2.8.0
torch>=1.11.0

# 可视化
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.6.0
folium>=0.12.0

# HTTP客户端
requests>=2.27.0
aiohttp>=3.8.0

# 工具库
python-dotenv>=0.19.0
pyyaml>=6.0
click>=8.0.0
tqdm>=4.62.0
loguru>=0.6.0

# 测试
pytest>=7.0.0
pytest-cov>=3.0.0
pytest-asyncio>=0.18.0

# 文档
sphinx>=4.5.0
sphinx-rtd-theme>=1.0.0
